LayoutTests/platform/chromium-win/media/video-dom-src-expected.txt

EXPECTED (video.currentSrc == '') OK
EXPECTED (video.currentSrc == '') OK
EVENT(canplaythrough)
EXPECTED (relativeURL(video.currentSrc) == 'content/test.ogv') OK
END OF TEST
